Prevent HTTP 414 Request-URI Too Large responses
The server refuses batch requests with large data sets.
'listBatchSubscribe', 'listBatchUnsubscribe', and 'listSubscribe' are
improperly converted POST requests to GET requests with long query
strings. To prevent this, the 'type' attribute of any cfhttpparam inside
a loop was set to 'formfield'.

E.g.: The server responds with 'HTTP 414: Request-URI Too Large' when
calling 'listBatchSubscribe' for 700 records.